moved position of sanitize_customFlavour
