+++ src/util/index.ts
@@ -4,4 +4,3 @@ export * from './dump';
 export * from './articleRenderers';
 export * from './const';
 export * from './mw-api';
-export * from './customFlavor';
