+++ test/unit/util.test.ts
@@ -3,7 +3,8 @@ import test from 'blue-tape';
 import { URL } from 'url';
 import fs from 'fs';
 import pathParser from 'path';
-import { encodeArticleIdForZimHtmlUrl, interpolateTranslationString, getFullUrl, getCustomFlavorPath } from 'src/util';
+import { sanitize_customFlavour } from 'src/sanitize-argument';
+import { encodeArticleIdForZimHtmlUrl, interpolateTranslationString, getFullUrl } from 'src/util';
 import { testHtmlRewritingE2e } from 'test/util';
 
 test('util -> interpolateTranslationString', async (t) => {
@@ -73,39 +74,37 @@ test('Custom flavour path', async(t) => {
     // checks in current working directory.
     let createStream = fs.createWriteStream(pathParser.resolve(process.cwd(), 'testCustomFlavour.js'));
 
-    t.equal(getCustomFlavorPath('testCustomFlavour.js'), pathParser.resolve(process.cwd(), 'testCustomFlavour.js'),
+    t.equal(sanitize_customFlavour('testCustomFlavour.js'), pathParser.resolve(process.cwd(), 'testCustomFlavour.js'),
         'Custom flavour in working directory.');
 
     // checks when current and working directory have file with same name(preference to working directory).
-    createStream = fs.createWriteStream(pathParser.resolve(process.cwd(), 'extensions/testCustomFlavour.js'));
+    createStream = fs.createWriteStream(pathParser.resolve(__dirname, '../../extensions/testCustomFlavour.js'));
 
-    t.equal(getCustomFlavorPath('testCustomFlavour'), pathParser.resolve(process.cwd(), 'testCustomFlavour.js'),
+    t.equal(sanitize_customFlavour('testCustomFlavour'), pathParser.resolve(process.cwd(), 'testCustomFlavour.js'),
         'When file with same name exist in extensions directory.');
 
     createStream.end();
     fs.unlinkSync(pathParser.resolve(process.cwd(), 'testCustomFlavour.js'));
-    fs.unlinkSync(pathParser.resolve(process.cwd(), 'extensions/testCustomFlavour.js'));
 
     // checks in extension directory.
-    t.equal(getCustomFlavorPath('wiktionary_fr.js'), pathParser.resolve(process.cwd(), 'extensions/wiktionary_fr.js'),
+    t.equal(sanitize_customFlavour('wiktionary_fr.js'), pathParser.resolve(__dirname, '../../extensions/wiktionary_fr.js'),
         'Custom flavour in extensions directory.');
 
-    t.equal(getCustomFlavorPath('wiktionary_fr'), pathParser.resolve(process.cwd(), 'extensions/wiktionary_fr.js'),
+    t.equal(sanitize_customFlavour('wiktionary_fr'), pathParser.resolve(__dirname, '../../extensions/wiktionary_fr.js'),
         'Custom flavour in extension directory without js extension.');
 
     // checks in absolute path.
-    createStream = fs.createWriteStream(pathParser.resolve(process.cwd(), 'extensions/testCustomFlavour.js'));
 
-    t.equal(getCustomFlavorPath(pathParser.resolve(process.cwd(), 'extensions/testCustomFlavour.js')),
-        pathParser.resolve(process.cwd(), 'extensions/testCustomFlavour.js'),
+    t.equal(sanitize_customFlavour(pathParser.resolve(__dirname, '../../extensions/testCustomFlavour.js')),
+        pathParser.resolve(__dirname, '../../extensions/testCustomFlavour.js'),
         'Positive check with absolute path.');
 
-    t.equal(getCustomFlavorPath(pathParser.resolve(process.cwd(), 'extensions/negativeTest.js')),
+    t.equal(sanitize_customFlavour(pathParser.resolve(__dirname, '../../extensions/negativeTest.js')),
         null, 'Negative test for absolute path.');
 
     createStream.end();
-    fs.unlinkSync(pathParser.resolve(process.cwd(), 'extensions/testCustomFlavour.js'));
+    fs.unlinkSync(pathParser.resolve(__dirname, '../../extensions/testCustomFlavour.js'));
 
     // negative scenario
-    t.equal(getCustomFlavorPath('wrongCustomFlavour.js'), null, 'Returning null when file doesnt exist.')
+    t.equal(sanitize_customFlavour('wrongCustomFlavour.js'), null, 'Returning null when file doesnt exist.')
 })
